typeを各フォルダに分ける（feature-akapo_fb08）
