+++ front/components/posts/models/user.ts
@@ -1,0 +1,20 @@
+export interface User {
+  id: number;
+  name: string;
+  iconUrl: string;
+  signedUrl: string;
+}
+
+export interface NewUser {
+  loginId: string;
+  name: string;
+  password: string;
+}
+
+export interface UserProfile {
+  id: number;
+  name: string;
+  loginId: string;
+  iconUrl?: string;
+  signedUrl?: string;
+}
