+++ front/components/posts/userArticleList/index.tsx
@@ -1,7 +1,7 @@
 "use client";
 import Link from "next/link";
 import Photo from "@/components/shared/photo";
-import { UserArticleListProps } from "@/types";
+import { UserArticleListProps } from "@/components/posts/models/post";
 
 export default function UserArticleList({ data }: UserArticleListProps) {
   if (!data) {
