+++ front/components/posts/userPage/index.tsx
@@ -1,6 +1,6 @@
 "use client";
 import { useEffect, useState } from "react";
-import { Post } from "@/types";
+import { Post } from "@/components/posts/models/post";
 import Image from "next/image";
 import { fetchUserPosts, fetchUserData } from "@/api/fetchData";
 import UserArticleList from "../userArticleList";
