+++ front/components/shared/types/auth/login.ts
@@ -1,0 +1,4 @@
+export interface NewLogin {
+  loginId: string;
+  password: string;
+}
