+++ front/components/shared/types/ui/photo.ts
@@ -1,0 +1,4 @@
+export interface PhotoProps {
+  src: string;
+  alt: string;
+}
