+++ front/components/shared/types/ui/selectBox.ts
@@ -1,0 +1,6 @@
+export interface SelectBoxProps {
+  options: { value: string; label: string }[];
+  value: string | string[];
+  onChange: (value: string | string[]) => void;
+  multiple?: boolean;
+}
