+++ front/components/shared/types/ui/tagList.ts
@@ -1,0 +1,3 @@
+export interface TagListProps {
+  Categories?: Array<{ name: string }>;
+}
